Intro to Conditional System
Started work on a Conditional system for scene transitions, or for anything else, that works smoothly from the inspector and less from script. It allows multiple conditions and value points to be set within a list, and shows update references for those conditions. Added in a condition trigger system, and so far have setup the system to work with a pickup that can set or add to values on a condition list. A condition list is already present by default on a scene transition point, and will automatically allow access to a scene change if all conditions are met, or the list is empty. Started implementing conditions within viking village 3D that are required to access the ship scene. Removed extra showcase test scenes.