3D Village Changes and Scene Layout from - 2DMines to 3D Ship
Changed Tafl boards over to pieces to collect, and made changes to Water that corrected the shader issues. (Mesh was not mid-poly from blender, but Unity's default plane, which was why the waves did not seem to smooth out). Correct spawn point and 2 scene transitions in 2D mine, as the doorways needed to move with back wall and caused misalignment with colliders. Laid Scenes out to correct chain through Ice Cave 2D, Ice Cave 3d, Ship 3D - w/ the ability to go back and forth between 2D and 3D ship, but not return to Cave