Updates to Dwarf1 for Demonstration
Notes: "Changed the old enemy-specific scripts to not reference the player at all and to use the function Enemy.DetectPlayer() and the TriggerDetector2D/3D component on the sound/attack triggers to detect the player and send the reference to the Enemy script. I've gone ahead and adjusted that script.

Next is that the enemy entered attack mode but didn't attack the player, which was then caused by the melee distance being too low; the enemy was trying to get closer to the player to attack then it could, so I've upped the distance on the 1st dwarf from 4 to 10

Also, don't include the Animator in the list of components to disable in the Health component; it's still needed for the death animation"