Fix for 3d village
LocalGameManager for 2DMine was deactivated, and the player character was directly added and saved again. The camera for the Kubb game was left on for testing, but by the time you turn it off in editor, the player was being killed in the 2dmine, and triggering errors that stop most processes.