Fates Quiz
When interacting with the fates. If the player does not have the mirror they will inform the player to come back when they have it.
If they have the mirror the quiz will start. If they pass the quiz they will be loaded into the Medusa fight. If they fail the quiz they will be loaded into the medusa fighter but will loose health first.