2D Player Respawn, Falling Scrolls, Enemies
The 2D player now respawns similarly to the 3D player. Removed unnecessary update code.
I added a collider to the Greek Library 2D scene for the falling scrolls.
The falling scrolls no longer run a timer on update. They now check on trigger enter for the collider and are destroyed at this time.
I fixed an issue with the enemy scripts not getting the local game manager on start.
The 2D mouse will no longer drop scrolls if the player is not in its scene.