3D Boar & Other Errors
TwoDLever.cs: Added a check for the bool data. It will no longer give an exception error. It will log a warning if the data is missing.
PedastalsPuzzleManager.cs: Commented out the DontDestroyOnLoad bit as it is not a root object and there for cannot use this code.
Attack3D.cs: Added a log warning if the enemy health is null.
GreekLabyrinth_3D.unity: Backed the NavMesh so that the enemies can move around and remove all the errors.
Boar3D.cs: Changed the Defend function to call an IEnumerator to allow the enemy to exit defense state. This might have to be applied to every 3D enemy script.
HealingPotion.cs: Removed debug logs.
Health.cs: Added log warning if the enemy script is missing or the animator is missing.