Greek Labyrinth
TwoDBow: The bow can now be acquired by the player. Once the pedestal puzzle is completed the bow will appear and a pop up will display letting the player now to pick it up. Picking it up will update the game managers game data.
TwoDMirror: The mirror can now be acquired by the player. Once the lever puzzle is completed the mirror will appear and a pop up will display letting the player now to pick it up. Picking it up will update the game managers game data.
ElementalPuzzleItem: Removed code in the update loop. All code now runs on trigger enter.
LeverPuzzle: Optimized code to prevent code from running when not needed.
PedastalsPuzzle: Removed some code from the update loop. Simplified the switch statements.
PedastalsPuzzleManager: Removed references to the DoorCode script and replaced it with SceneTransitionPoint2D
WeaponSwitcher: The player can no longer equip the mirror if they do not have it.
GreekLabyrinth_3D: Added the orange diamond gem. The player can now access the GreekLabyrinth_2D_Pedastals scene.
QuizManager: Now checks the game data to see if the player has the mirror.
GameData: Added a ref to the bow.
GameManager: Added a helper function for testing that runs when the F1 key is pressed. This will transition the player to the next scene.