Greek Cavern 2D Lever Puzzle and Owl Companion
The player can now enter the Greek Cavern 2D Lever Puzzle.
The player can now activate the levers in the following sequence (on, off, on, off ...)
Once activated properly the owl is obtained and saved to json.
The owl will appear on all players that have this companion after this point and will the follow the player.