Bow, Mirror, Lever Puzzles, Companion
Fixed how the bow and mirror were being saved. There was a bug introduced with may changes to the save system.
Combined the owl and mirror puzzle manager scripts. Updated their scenes to reflect this.
Fixed a bug with the companion controller that would not update the player var on the companion in every instance.