Weapon Switcher
Fixed the mirror displaying on load in some instances.
Fixed the bow being displayed if the player has not yet obtained it.
Fixed the door not appearing after defeating medusa.
Added a collider in the 3D Greek Labyrinth scene preventing the player from climbing the stairs at the end and falling out of the map.