Build Bugs
Fixed a bug that would crash the build if the player tried to continue without a save file.
Fixed a bug that would crash the build if the player returns to the main menu and tries to start a new game or continue.
Changed the save data file path to not save to the game files directory.