3D Greek Library & 3D Mouse
I Moved all of the mouse sfx into there own folder inside of the SFX folder.
I adjusted the trigger sizes of the 3D health potion, the 3D Scrolls and 3D Levers.
This fixes the odd behavior of grabbed items when looking at one of these objects.
Enemies that use the enemySimpleAI will have a hurtSFX, attackSFX and a dieSFX that will need to be set.