Water Fountain & Greek Athens 2D
Added a fountain sfx to the water fountain in the Greek Athens 2D level.
As the player gets close to the volume increases and decreases as the player moves away from it.
If the player is close enough to it a interact prompt will appear.
If the player is damaged they can heal else it will say Full Health.