Bug Fixes and SFXs
Fixed a bug with the push/pull block in 2D Greek Library. The push/pull sfx would player even if the player was not moving.
Fixed a bug with the subliminal mechanic. When picking up items, if they have a sfx it now plays.
Set all sfxs to not play on awake.
Added the lever interact sfx to all levers in the Greek levels.