Torches, Lighting and SFXs
Removed directional lighting from the 2D labyrinth levels.
Added a fire particle effect to the torches in the 3D labyrinth level.
Added a particle light to each of the torches in the 3D labyrinth level.
Added a 3D torch SFXto each of the torches in the 3D labyrinth level.
Added a interact prompt to the elemental items in the 3D labyrinth level.
Added a SFX on picking up the elemental items in the 3D labyrinth level.