2D Animations
Organized all of the 2D Greek animations.
Recreated all of the 2D Greek animations and animator controllers so that they are all the same.
Checked all 2D Dying animations to make sure they are not looping.
Went through all of the 2D Greek levels and verified the animations are working.
Some animations are missing still. But the ones we have are working.