2D/3D Greek Library
Fixed a bug where the 2D mice would not drop the scrolls.
Fixed a warning message regarding the 2D mice.
Adjusted spacing of the 2D library's pedestal, door and lever. This fixes a bug where the doors interact UI would not appear in some instances.
Fixed a bug where the enter UI would not disappear after entering a 2D or 3D door.