2D Bow Attack
Simplified the Shoot2D script.
The 2D bow is now activated by pressing 1.
The 2D bow will not work if the game is paused, the player has not collected the bow or of the player is pushing or climbing.
Cleaned up the 2D arrow .png.
Fixed the 2D arrow box collider.
The 2D arrow will no longer collide with the player.
Fixed a bug that prevented the sword animation from playing after switching to the bow and back to the sword.
Fixed a bug with the 2D bird collider.