Bug Fixes
Made a build for testing. Found some bugs.
Added a scroll for when the player gets the owl to explain how to activate it.
Fixed a bug with the weapon switcher that prevented the player from switching in some instances.
Fixed a bug with the companion switcher that prevented the player from dismissing the companion.
Fixed a bug that prevented the 2D mice from dropping scrolls if they were assigned to do so.
Fixed a build error.