Bug Fixes & GreekCavern_2D_Levers
Fixed a bug that allowed the 2D player to attack with the sword and bow at the same time.
Fixed a bug with the torches in the GreekCavern_2D_Levers level. Torches no longer flicker.
Added a scroll to the GreekCavern_2D_Levers level.
Added more columns in GreekCavern_2D_Levers level for the levers to be on.