Bug Fixes and Improvements
Fixed a some tree placements in the Athens scene.
Reduced the owls damage to 1 from 3.
Added a hurt and die SFX to Medusa.
Added the hurt animation to Medusa.
Increased Medusa's SFX volume.
Increased Medusa's projectile speed.
Added lighting to the medusa scene.
Added fire effect to the braziers in the medusa scene with SFX and Lighting.